A bunch of stuff
Moved a lot into the movement component rather than the pawn, that way I can apply it to a random pawn to some extent. I also changed the vector thing, though I'm not sure that will carry over so someone else will have to check. FVector now has a function DistanceInDirection. It will give the length of the vector in the direction of another given vector. This can be positive or negative and I'm pretty damn sure my answer is right