Lost One suspicion adding
The Lost Ones will add suspicion. Not done with that stuff quite yet though. Can use a scalar for how perceptive it is, works with distance, but make sure it's in centimeters. Also there's a little bug with going down stairs Can probably just increase step height but I'm not sure that 's the best choice. Probably better to just use ramps for the colliders on stairs